# Optimize `Array#*`